Moved the last components.
