+++ frontend/lib/app_component.dart
@@ -3,12 +3,12 @@ import 'package:angular2/platform/common.dart';
 import 'package:angular2/router.dart';
 
 import 'package:frontend/components/event_list_component/event_list_component.dart';
-import 'event_detail_component.dart';
+import 'package:frontend/components/event_detail_component/event_detail_component.dart';
+import 'package:frontend/components/navigation_bar_component/navigation_bar_component.dart';
 import 'package:frontend/store/regatta_store.dart';
 import 'package:frontend/services/event_service.dart';
 import 'package:frontend/services/sailing_club_service.dart';
 import 'package:frontend/services/person_service.dart';
-import 'package:frontend/navigation_bar_component.dart';
 
 @Component(selector: 'my-app', templateUrl: 'app_component.html', directives: const [
   ROUTER_DIRECTIVES,
