+++ frontend/lib/components/event_detail_component/event_detail_component.html
