+++ frontend/lib/components/navigation_bar_component/navigation_bar_component.dart
